🚨 Fix: Apply 기능 제거, 우측 컨텐츠 PostStickyCard.vue로 분리리
